+++ scglobals/static/tx/src/app/models/shipment.js
@@ -3,6 +3,6 @@ import TablesDefaultModel from './tables-default';
 
 export default TablesDefaultModel.extend({
   defaults: {
-    transactions: []
+    // transactions: []
   }
 });