+++ scglobals/static/tx/src/app/models/tables-default.js
@@ -1,6 +1,10 @@
 import Backbone from 'backbone';
+import DB from './../services/db';
 
 export default Backbone.Model.extend({
+  initialize: function(){
+    this.db = new DB();
+  },
   defaults: function () {
     return {
       created: new Date(),
@@ -8,5 +12,11 @@ export default Backbone.Model.extend({
       creator: '',
       updated_by: ''
     };
+  },
+  fetch: function(){
+    return this.db.get(this.id).then((results)=>{
+      console.log(results);
+      return this.set(results);
+    });
   }
 });